refactor: rename proifle-area to user-menu and move to user components
